Add Docstrings and Refactors for Readability
